# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hotspot API",
    "description": "Community driven motorcycle parking suggestions for Victoria, Australia",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://hotspot.sknk.ws",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/risk/top": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Get risk rankings",
        "description": "Paginated list of suburbs or LGAs ranked by safety scores with search and sorting",
        "operationId": "risk_top_api_v1_risk_top_get",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(postcode|lga)$",
              "description": "Group by postcode or lga",
              "default": "postcode",
              "title": "Scope"
            },
            "description": "Group by postcode or lga"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Results per page",
              "default": 20,
              "title": "Itemsperpage"
            },
            "description": "Results per page"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Column to sort by",
              "title": "Sortby"
            },
            "description": "Column to sort by"
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "Sort direction",
              "default": "desc",
              "title": "Sortorder"
            },
            "description": "Sort direction"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Search filter text",
              "title": "Search"
            },
            "description": "Search filter text"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated risk data with total count",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRiskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedRiskResponse": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items",
            "description": "Risk ranking items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of items"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "PaginatedRiskResponse",
        "description": "Paginated risk ranking response"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}